:recycle: Move and rename getClassList & getClassMap
